Merge trytond_nereid into nereid
Based on issue #30 the package was separated into the python module
nereid and the tryton module trytond_nereid. In retrospective this was a
bad idea and the mainteinance of the module became a nightmare with
most changes requiring two commits on separate repositories.

This patch merges the trytond_nereid package back into nereid.